# [CH] Improve master_commit_red_avg query perf