--- a/torchci/clickhouse_queries/master_commit_red_avg/params.json
+++ b/torchci/clickhouse_queries/master_commit_red_avg/params.json
@@ -4,5 +4,18 @@
     "stopTime": "DateTime64(3)",
     "workflowNames": "Array(String)"
   },
-  "tests": []
+  "tests": [
+    {
+      "startTime": "2025-03-18T21:09:47.987",
+      "stopTime": "2025-03-25T21:09:47.987",
+      "workflowNames": ["lint", "pull", "trunk"],
+      "_comment": "1 week"
+    },
+    {
+      "startTime": "2025-02-18T21:09:47.987",
+      "stopTime": "2025-03-18T21:09:47.987",
+      "workflowNames": ["lint", "pull", "trunk"],
+      "_comment": "1 month"
+    }
+  ]
 }